requirements-api.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiohttp==3.9.0
websockets==12.0
python-multipart==0.0.6
Pillow==10.1.0
persist-queue==1.0.0